split barcode scanner code into multiple files
